move the component closer to the screen
